% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mentions.R
\name{count_mentions_in_dataframe}
\alias{count_mentions_in_dataframe}
\title{Count number of occurrences key words in a text column of a data frame}
\usage{
count_mentions_in_dataframe(
  dataframe,
  words,
  ignore_case = TRUE,
  word_boundaries = TRUE,
  return_all_rows = FALSE
)
}
\arguments{
\item{dataframe}{data frame, with a column called \code{"text"}}

\item{words}{vector of key words to search for}

\item{ignore_case}{logical, should case sensitivity be considered?}

\item{word_boundaries}{logical, should word boundaries be considered?}

\item{return_all_rows}{logical, return all rows, or only rows with 1+
instance? It's considerably faster to only return rows with 1+ instance for
large data frames and/or infrequent words.}
}
\description{
Count number of occurrences key words in a text column of a data frame
}
