% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_counts.R
\name{geom_lollipop}
\alias{geom_lollipop}
\alias{geom_h_lollipop}
\alias{geom_v_lollipop}
\title{Lollipop plot}
\usage{
geom_h_lollipop(mapping = NULL, data = NULL, ..., .id = character(0L))

geom_v_lollipop(mapping = NULL, data = NULL, ..., .id = character(0L))
}
\description{
Lollipop plots are (visually) lighter-weight versions of bar plots. This
ggpacket is composed of geom_point and geom_segment.
}
\section{Functions}{
\itemize{
\item \code{geom_h_lollipop()}: Lollipop plot with horizontal lines

\item \code{geom_v_lollipop()}: Lollipop plot with vertical lines

}}
