% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtextcounts-package.R
\docType{package}
\name{ggtextcounts-package}
\alias{ggtextcounts}
\alias{ggtextcounts-package}
\title{ggtextcounts: Tools for counting & plotting instances of a word or string in a data frame}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Alice Circe Malone \email{circemalone@gmail.com}

}
\keyword{internal}
