% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mentions.R
\name{padb}
\alias{padb}
\title{Pad word so regex searches for word boundaries}
\usage{
padb(word)
}
\arguments{
\item{word}{word (or string) to pad}
}
\description{
Pad word so regex searches for word boundaries
}
